refactor: Extract KuksaBridge into kuksa_bridge.hpp
