+++ src/ADAS-Manager/drive_mode.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+enum class DriveMode { MANUAL, AUTONOMOUS };
